How to Create a Simple Dialog Box with Angular Material
You don't have to be a CSS master to build nice dialog boxes.

In a web application, dialog boxes are convenient tools to interact
with users. It can attract users' attention if used properly and will
make your webpage look neat.

It needs some advanced CSS knowledge if you want to build a dialog
box from scratch. However, using a framework like Angular, creating
dialog boxes is a very simple task, even for someone who is not an
expert in CSS. In this post, we will demonstrate how to create a
simple dialog box with Angular. All the details as well as some
caveats are covered which can be helpful for your work in practice.

Reference:
  - https://medium.com/p/8cd68440ddd1